howto : proposal on "how to" section for documentation
Please shoot at this so I get a feeling how best to contribute my
first steps to other people wanting to contribute. You could argue that
this is maybe a little bit toooooo simple. Let's take this as something
to form our opinions and discuss :)
I don't know if the files etc are in the right place. I favor another
(fold down) structure in the side panel (but don't know how to do so,
or if this is the appropriate location) like:
HOWTO
--first contribution
  --about
  --Git
  --Commits
  --Rebase and Push
  --Pull Request